+++ web/src/elements/forms/SearchSelect/index.ts
@@ -1,0 +1,4 @@
+import { SearchSelect } from "./ak-search-select";
+
+export { SearchSelect };
+export default SearchSelect;
